+++ client.src/core/views/modal-wrapper/index.js
