+++ client.src/core/views/overlay/index.js
